refactor token generator and associated tests
made encodeHashData private
removed token generator test fixture
renamed token generator test to match naming conventions
fixed various typos
WIP - symfonyized tests
