+++ ResetPasswordBundle/src/Generator/ResetPasswordTokenGenerator.php
@@ -23,7 +23,7 @@ final class ResetPasswordTokenGenerator
     }
 
     //@todo make me private | fix tests for private
-    protected function encodeHashData(\DateTimeInterface $expiresAt, string $verifier, string $userId): string
+    private function encodeHashData(\DateTimeInterface $expiresAt, string $verifier, string $userId): string
     {
         return \json_encode([
             $verifier,
@@ -31,67 +31,4 @@ final class ResetPasswordTokenGenerator
             $expiresAt->format('Y-m-d\TH:i:s')
         ]);
     }
-
-//    /** @throws \Throwable */
-//    private function isEmpty(string $value): void
-//    {
-//        if (empty($value)) {
-//            throw TokenException::getIsEmpty();
-//        }
-//    }
-//
-//    /** @throws TokenException */
-//    private function isExpireValid(\DateTimeImmutable $expire): void
-//    {
-//        $time = $expire->getTimestamp();
-//
-//        if ($time <= time()) {
-//            throw TokenException::getInvalidTokenExpire();
-//        }
-//    }
-//
-//    protected function generateHash(
-//        string $signingKey,
-//        \DateTimeImmutable $expiresAt,
-//        string $verifier,
-//        string $userId
-//    ): string {
-//
-//        return \hash_hmac(
-//            self::HMAC_HASH_ALGO,
-//            $this->encodeHashData($expiresAt, $verifier, $userId),
-//            $signingKey,
-//            false
-//        );
-//    }
-
-//    /**
-//     * @TODO get me outta here
-//     * Original credit to Laravel's Str::random() method.
-//     */
-//    public function getRandomAlphaNumStr(int $length): string
-//    {
-//        $string = '';
-//
-//        while (($len = strlen($string)) < $length) {
-//            $size = $length - $len;
-//
-//            $bytes = \random_bytes($size);
-////            $bytes = $this->getRandomBytes($size);
-//
-//            $string .= substr(
-//                str_replace(['/', '+', '='], '', base64_encode($bytes)), 0, $size);
-//        }
-//
-//        return $string;
-//    }
-
-//    protected function getRandomBytes(int $size): string
-//    {
-//        try {
-//            return \random_bytes($size);
-//        } catch (\Error $exception) {
-//            throw TokenException::getBadBytes();
-//        }
-//    }
 }
